some comments and stuff

+++ src/main/java/com/codeup/adlister/controllers/AdminServlet.java
@@ -14,9 +14,10 @@ import java.io.IOException;
 public class AdminServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
         User user = (User)request.getSession().getAttribute("user");
         if (!user.getUsername().contains("admin")){
-            response.sendRedirect("/profile");
+            response.sendRedirect("/login");
         }
         request.setAttribute("users", DaoFactory.getUsersDao().allUsers());
         request.getRequestDispatcher("/WEB-INF/admin.jsp").forward(request, response);
@@ -25,10 +26,19 @@ public class AdminServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+//      if we've gotten here, we're either going to delete a user or make them an admin
+        if(request.getParameter("id")!=null){
         long user_id = Long.parseLong(request.getParameter("id"));
         DaoFactory.getAdsDao().deleteFromUser(user_id);
        DaoFactory.getUsersDao().deleteUser(user_id);
+            response.sendRedirect("/profile");
 
-       response.sendRedirect("/admin");
+        } else {
+            long user_id = Long.parseLong(request.getParameter("make-admin"));
+            DaoFactory.getUsersDao().makeAdmin(DaoFactory.getUsersDao().findByUserId(user_id));
+            response.sendRedirect("/profile");
+
+        }
+
     }
 }
